+++ packages/tendermint-rpc/src/encodings.ts
@@ -1,12 +1,7 @@
 import { fromBase64, fromHex, fromRfc3339, toBase64, toHex, toUtf8 } from "@cosmjs/encoding";
 import { Int53 } from "@cosmjs/math";
-import { As } from "type-tagger";
 
 import { BlockId, ReadonlyDateWithNanoseconds, Version } from "./responses";
-
-export type Base64String = string & As<"base64">;
-export type HexString = string & As<"hex">;
-export type IntegerString = string & As<"integer">;
 
 /**
  * A runtime checker that ensures a given value is set (i.e. not undefined or null)
@@ -150,22 +145,22 @@ export function dictionaryToStringMap(obj: Record<string, unknown>): Map<string,
 }
 
 export class Integer {
-  public static parse(input: IntegerString | number): number {
+  public static parse(input: string | number): number {
     const asInt = typeof input === "number" ? new Int53(input) : Int53.fromString(input);
     return asInt.toNumber();
   }
 
-  public static encode(num: number): IntegerString {
-    return new Int53(num).toString() as IntegerString;
+  public static encode(num: number): string {
+    return new Int53(num).toString();
   }
 }
 
 export class Base64 {
-  public static encode(data: Uint8Array): Base64String {
-    return toBase64(data) as Base64String;
+  public static encode(data: Uint8Array): string {
+    return toBase64(data);
   }
 
-  public static decode(base64String: Base64String): Uint8Array {
+  public static decode(base64String: string): Uint8Array {
     return fromBase64(base64String);
   }
 }
@@ -187,11 +182,11 @@ export class DateTime {
 }
 
 export class Hex {
-  public static encode(data: Uint8Array): HexString {
-    return toHex(data) as HexString;
+  public static encode(data: Uint8Array): string {
+    return toHex(data);
   }
 
-  public static decode(hexString: HexString): Uint8Array {
+  public static decode(hexString: string): Uint8Array {
     return fromHex(hexString);
   }
 }